+++ src/utils/utils.ts
@@ -1,5 +1,7 @@
 import type { Endpoint } from '../interface/observation';
 import type { ComponentDefinitionsBase } from '../interface/application';
+import type { LoginUserInfo } from '../interface/user';
+import _ from 'lodash';
 
 type SelectGroupType = {
   label: string;
@@ -141,4 +143,16 @@ export function getSelectLabel(
 
 export function getMatchParamObj(match: { params: any }, name: string) {
   return match.params && match.params[name];
+}
+
+export function isAdminUserCheck(userInfo: LoginUserInfo | undefined) {
+  const platformPermissions = userInfo?.platformPermissions || [];
+  const findAdminUser = _.find(platformPermissions, (item) => {
+    return item.name === 'admin';
+  });
+  if (findAdminUser) {
+    return true;
+  } else {
+    return false;
+  }
 }
